Greatly improve the utility of Motor board exceptions
Have each of them include the serial number of the board to which
they relate, thus making it easier for the user to tell which board
is at fault if they have more than one connected.

Add a hint as to the typical cause of failing to read the firmware
from the board in the exception message (this has been asked about
on the forums 3 times already this year that I recall).

Change-Id: Ibb264b5264673f5097c007c54ee280f5c90464f6