Add support for querying power board output states
This is implemented in the same way as we have done for servo angles
as I'm not aware of a mechanism to query them directly. Also move
index checking to a helper and use the more specific IndexError.

Patch set 2: fix initial state values (based on the fact that we
             explicitly set them all to True in Robot._init_power
             with a comment about bringing them up)

Change-Id: Iaa33011faedb7a6fbffae64aa78af0389bc2aaaa